% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datot}
\alias{datot}
\title{Housing Return on Equity(ROE) with WalkScore}
\format{A data frame with 720 rows and 8 variables:
\describe{
  \item{formatted_address}{full address with city and zipcode}
  \item{beds}{bed information within the unit}
  \item{baths}{bath information within the unit}
  \item{postal_code}{zipcode accompanied with address}
  \item{latitude}{geodata of location}
  \item{longitude}{geodata of location}
  \item{roe}{return on equity for each unit}
  \item{walk_score}{walk score for each unit}
}}
\usage{
datot
}
\description{
A dataset that contains the calculation of ROE and walk score
combined by zipcode for every housing in Manhattan.
}
\keyword{datasets}
