% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datr}
\alias{datr}
\title{Rental Data with other housing information}
\format{A data frame with 12082 rows and 9 variables:
\describe{
  \item{street}{address with unit info}
  \item{rent}{rent with punctuation mark}
  \item{beds}{bed information within the unit}
  \item{baths}{bath information within the unit}
  \item{ftsq}{size of the unit in ft square}
  \item{address}{address without unit info}
  \item{unit}{unit info}
  \item{rent_p}{numeric version of rent}
  \item{area_ft}{numeric version of area ft}
}}
\usage{
datr
}
\description{
A dataset that collect Manhattan housing for rental, with both
format of text and number for unit info.
}
\keyword{datasets}
