% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ws_df}
\alias{ws_df}
\title{Walk Score Data with zipcode}
\format{A data frame with 42 rows and 2 variables:
\describe{
  \item{postal_code}{zipcode}
  \item{walk_score}{walk score for each unit}
}

For further details, see \url{https://www.walkscore.com/professional/}}
\usage{
ws_df
}
\description{
A dataset include walk score and corresponding zipcode.
}
\keyword{datasets}
